made rating modal a component
